requirements.txt
aiofiles
annotated-types
anyio
bcrypt
certifi
click
colorama
dnspython
ecdsa
email-validator==2.0.0.post2
fastapi
greenlet
h11
httpcore
httptools
httpx
idna
inflect
itsdangerous
Jinja2
mangum
MarkupSafe
orjson
passlib
Pillow
psycopg2-binary
pyasn1
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
python-dotenv
python-jose
python-multipart
PyYAML
rsa
six
sniffio
sqlacodegen==2.3.0.post1
SQLAlchemy
starlette
typing_extensions
ujson
uvicorn
watchfiles
websockets
python-dateutil